get back the devToolAdapter
